move value stuff to separate file and add more From impls
